requirements.txt
asgiref==3.5.2
Django==3.2.14
django-cors-headers==3.13.0
django-model-utils==4.2.0
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
phonenumbers==8.12.52
Pillow==9.2.0
pycodestyle==2.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
